Refactor createSeekGraphMiddleware
- Extract a `createAgent` helper for reuse across tests
- Move constants out of `seekGraph`
